util/ExportToCsv: Only duplicate media on export when they are unique
While exporting, if the destination folder already contains a file with the
same name, compute the MD5 hash of the files and compare them. If
same, skip copying; else append a suffix to differentiate the files.
For better performance, cache the already computed hash in a HashMap with
fileName-hashValue as key-value pair.

Fix #77