Simplify form metadata structure to deal with form dirs
We need to have FormMetadata objects in memory that have no storage location info in the interim between getting them in a list of remote forms and pulling them. The storage location will be provided after puling them. That's why the two commands that dealt with updating metadata as pulled now are generic upsert commands.

They need to be upsert commands as opposed to insert commands because a form can be pulled twice.

The update last exported submission command, on the other hand, must me an update that needs to fail if the metadata is unknown. If there's no metadata, the form effectively hasn't been pulled, which should be an impossible scenario.