Ensure that the metadata port and form caches are properly flushed
They have to be flushed whenever the user changes the storage location. Otherwise, formdefs or form metadata objects could point to old locations if it happens that a form is present both in the previous and current sds.

Use fresh FormStatus list to avoid problems when changing sd locations

If a form happens to be both in the previous and current locations, we need to be sure we have a fresh FormStatus with a fresh form definition pointing to the new sd location