Store relative paths only in database
We need to store relative paths (form and submission metadata) to make the sd portable. The adapter now resolves and relativizes paths using the workspace, which forced us to separate a dependency injection container from the workspace definition due to cyclic dependencies (the metadata adapters depend on the workspace, and the workspace required the adapters to be created).

Now, there's a container object with the dependencies that implements the program's lifecycle.