Fix pull CLI ops to avoid losing prefs for the GUI
Somehow, setting the sd in the pull CLI ops was losing prefs for the GUI, and users were asked to set it again in the next GUI launch.

Also, setting the sd in CLI ops is not required (downstream code gets the briefcaseDir directly), so it's a safe change to do.