A solution for Bug #798 (#851)
The bug was caused by the fact that the Source/Target panel has no information that the forms are being loaded from the server. Making a call to setWorking() in the current view makes the buttons in the Source/Target panel to go disabled until the unsetWorking method is call. I think this fixes the bug as the user won be able to click on reset (or reload) during the time the forms are being loaded, and also the user will have good feedback.